refactor(protocol): extract wire contract to top-level mypal_protocol package
