+++ mypalclara/adapters/cli/gateway_client.py
@@ -19,7 +19,7 @@ from rich.text import Text
 from mypalclara.adapters.base import GatewayClient
 from mypalclara.adapters.manifest import AdapterManifest, adapter
 from mypalclara.config.logging import get_logger
-from mypalclara.gateway.protocol import ChannelInfo, UserInfo
+from mypal_protocol import ChannelInfo, UserInfo
 
 logger = get_logger("adapters.cli.gateway")
