+++ mypalclara/adapters/teams/gateway_client.py
@@ -20,7 +20,7 @@ from mypalclara.adapters.manifest import AdapterManifest, adapter
 from mypalclara.adapters.teams.graph_client import GraphClient, get_graph_client
 from mypalclara.adapters.teams.message_builder import AdaptiveCardBuilder
 from mypalclara.config.logging import get_logger
-from mypalclara.gateway.protocol import ChannelInfo, UserInfo
+from mypal_protocol import ChannelInfo, UserInfo
 
 if TYPE_CHECKING:
     from mypalclara.adapters.teams.bot import TeamsBot
